requirements.txt
djangorestframework
psycopg2==2.8.6
django==3.1.7
pytest
pytest-django
